pyproject.toml
[project]
name = "notas-dian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "mysql-connector-python>=9.3.0",
    "openpyxl>=3.1.5",
    "ttkbootstrap>=1.13.11",
]
